+++ csrc/ops/utils.cpp
@@ -422,8 +422,8 @@ Val* getMaximumValue(DataType v) {
 }
 
 std::vector<unsigned int> canonicalizeAxes(
-    const std::vector<int>& axes,
-    size_t ndims) {
+    const std::vector<int64_t>& axes,
+    int64_t ndims) {
   std::vector<unsigned int> uint_axes;
   uint_axes.reserve(axes.size());
   std::transform(
