+++ csrc/scheduler/mma_utils.h
@@ -53,7 +53,7 @@ NVF_API void scheduleWarpTileWithNoReduction(
 //! Eg.
 //!  A[B,I0,I1,I2] -> makeTile({1,2,3})
 //! Gives A[B, I0o, I1o, I2o, I0i(1), I1i(2), I2i(3)]
-void makeTile(TensorView* tv, std::vector<int> tile_sizes);
+void makeTile(TensorView* tv, std::vector<int64_t> tile_sizes);
 
 //! Order the inner tile dimensions as the original order in
 //!  root domain. Also putting broadcast domains on the left.
@@ -199,8 +199,8 @@ class WarpMmaSwizzler {
 
 void checkDimSize(
     TensorView* tv,
-    std::vector<int> axis,
-    std::vector<int> expect);
+    std::vector<int64_t> axis,
+    std::vector<int64_t> expect);
 
 //! A constant with minimum number of fusion inputs that could be MMA inputs.
 //!  TODO: update for square matmuls where both inputs are the same tensor
